add-package

+++ src/modules/resume/resume.service.ts
@@ -1,4 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Resume } from './entities/resume.entity';
 import { Repository } from 'typeorm';
@@ -10,9 +14,13 @@ export class ResumeService {
     private readonly resumeRepository: Repository<Resume>,
   ) {}
   async create(candidateId: number, name?: string) {
+    const existingResumes = await this.resumeRepository.find({
+      where: { candidate: { id: candidateId } },
+    });
     return this.resumeRepository.save({
       candidate: { id: candidateId },
       name: name ? name : 'Hồ sơ ' + candidateId,
+      isDefault: existingResumes.length === 0,
     });
   }
   async getAll(candidateId: number) {
@@ -49,6 +57,38 @@ export class ResumeService {
     if (!resume) {
       throw new Error('Hồ sơ không tồn tại hoặc bạn không có quyền truy cập');
     }
+    if (resume.isDefault) {
+      throw new BadRequestException('Không thể xóa hồ sơ mặc định.');
+    }
     return this.resumeRepository.remove(resume);
+  }
+
+  async toggleDefault(candidateId: number, resumeId: number) {
+    const resume = await this.validateMe(candidateId, resumeId);
+    if (!resume) {
+      throw new UnauthorizedException(
+        'Hồ sơ không tồn tại hoặc bạn không có quyền truy cập',
+      );
+    }
+    const listResumes = await this.resumeRepository.find({
+      where: {
+        candidate: { id: +candidateId },
+      },
+    });
+    if (listResumes.length === 0) {
+      throw new BadRequestException(
+        'Bạn không có hồ sơ nào để đặt làm mặc định',
+      );
+    }
+    if (listResumes.length === 1 && listResumes[0].isDefault) {
+      listResumes[0].isDefault = true;
+      throw new BadRequestException('Bắt buộc có ít nhất một hồ sơ mặc định');
+    }
+    listResumes.map((item) => {
+      item.isDefault = false;
+      item.isDefault = item.id === resumeId;
+      return item;
+    });
+    await this.resumeRepository.save(listResumes);
   }
 }
